+++ lib/Service/FolderService.php
@@ -135,38 +135,45 @@ class FolderService {
 	}
 
 	/**
-	 * @param $userId
-	 * @param $folderId
+	 * @param string $userId
+	 * @param int $folderId
 	 * @throws DoesNotExistException
 	 * @throws MultipleObjectsReturnedException
 	 * @throws UnsupportedOperation
+	 * @throws Exception
 	 */
-	public function deleteSharedFolderOrFolder($userId, $folderId): void {
-		/**
-		 * @var $folder Folder
-		 */
+	public function deleteSharedFolderOrFolder(string $userId, int $folderId, bool $hardDelete): void {
 		$folder = $this->folderMapper->find($folderId);
 
 		if ($userId === null || $userId === $folder->getUserId()) {
-			$this->treeMapper->deleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
+			if ($hardDelete) {
+				$this->treeMapper->deleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
+			} else {
+				$this->treeMapper->softDeleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
+			}
 			return;
 		}
 
 		try {
 			// folder is shared folder
-			/**
-			 * @var $sharedFolder SharedFolder
-			 */
 			$sharedFolder = $this->sharedFolderMapper->findByFolderAndUser($folder->getId(), $userId);
-			$this->treeMapper->deleteEntry(TreeMapper::TYPE_SHARE, $sharedFolder->getId());
+			if ($hardDelete) {
+				$this->treeMapper->deleteEntry(TreeMapper::TYPE_SHARE, $sharedFolder->getId());
+			} else {
+				$this->treeMapper->softDeleteEntry(TreeMapper::TYPE_SHARE, $sharedFolder->getId());
+			}
 			return;
 		} catch (DoesNotExistException $e) {
 			// noop
 		}
 
 		// folder is subfolder of share
-		$this->treeMapper->deleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
-		$this->folderMapper->delete($folder);
+		if ($hardDelete) {
+			$this->treeMapper->deleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
+			$this->folderMapper->delete($folder);
+		} else {
+			$this->treeMapper->softDeleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
+		}
 	}
 
 	/**
@@ -177,6 +184,31 @@ class FolderService {
 	 */
 	public function deleteShare($shareId): void {
 		$this->treeMapper->deleteShare($shareId);
+	}
+
+	/**
+	 * @throws UnsupportedOperation
+	 * @throws MultipleObjectsReturnedException
+	 * @throws DoesNotExistException|Exception
+	 */
+	public function undelete(?string $userId, int $folderId): void {
+		$folder = $this->folderMapper->find($folderId);
+		if ($userId === null || $userId === $folder->getUserId()) {
+			$this->treeMapper->softUndeleteEntry(TreeMapper::TYPE_FOLDER, $folderId);
+			return;
+		}
+
+		try {
+			// folder is shared folder
+			$sharedFolder = $this->sharedFolderMapper->findByFolderAndUser($folder->getId(), $userId);
+			$this->treeMapper->softUndeleteEntry(TreeMapper::TYPE_SHARE, $sharedFolder->getId());
+			return;
+		} catch (DoesNotExistException $e) {
+			// noop
+		}
+
+		// folder is subfolder of share
+		$this->treeMapper->softUndeleteEntry(TreeMapper::TYPE_FOLDER, $folder->getId());
 	}
 
 	/**
